feat: Add Firebase CLI installation to setup script
The setup script has been updated with a new function setup_firebase() that facilitates the installation of the Firebase CLI. This is achieved by running a new firebase.sh script, which downloads and executes the Firebase tool. The setup_firebase() function will now be executed as a part of the main setup process along with other existing setup functions.